+++ Source/Modules/Tests/Versioning/Gapotchenko.FX.Versioning.Semantic.Tests/SemanticVersionTests.Parse.cs
@@ -10,31 +10,6 @@ namespace Gapotchenko.FX.Versioning.Semantic.Tests;
 
 partial class SemanticVersionTests
 {
-    static IEnumerable<object[]> SemanticVersion_Parse_TestData_ValidInputs
-    {
-        get
-        {
-            yield return ["1.0.0"];
-            yield return ["1.0.0-alpha"];
-            yield return ["1.0.0-alpha.1"];
-            yield return ["1.0.0-alpha.beta"];
-        }
-    }
-
-    static IEnumerable<object[]> SemanticVersion_Parse_TestData_InvalidInputs
-    {
-        get
-        {
-            yield return [""];
-            yield return ["."];
-            yield return [","];
-            yield return ["1.abc.3"];
-            yield return ["123!"];
-            yield return ["1a:2b:3c"];
-            yield return ["123,abc,123"];
-        }
-    }
-
     #region Parse
 
     [TestMethod]
@@ -119,6 +94,35 @@ partial class SemanticVersionTests
 
         Assert.IsFalse(SemanticVersion.TryParse(input, out version));
         Assert.IsNull(version);
+    }
+
+    #endregion
+
+    #region Data
+
+    static IEnumerable<object[]> SemanticVersion_Parse_TestData_ValidInputs
+    {
+        get
+        {
+            yield return ["1.0.0"];
+            yield return ["1.0.0-alpha"];
+            yield return ["1.0.0-alpha.1"];
+            yield return ["1.0.0-alpha.beta"];
+        }
+    }
+
+    static IEnumerable<object[]> SemanticVersion_Parse_TestData_InvalidInputs
+    {
+        get
+        {
+            yield return [""];
+            yield return ["."];
+            yield return [","];
+            yield return ["1.abc.3"];
+            yield return ["123!"];
+            yield return ["1a:2b:3c"];
+            yield return ["123,abc,123"];
+        }
     }
 
     #endregion